Address PR feedback
Authored-by: Chris Hajas <chajas@pivotal.io>
